Remove ArgsForHost class
Just allow arrays of SimpleTypes/and ISerializable to be passed in
